Don't cut the value on `change`
Android does not support `keypress` which means when `change` runs, it will use the input and set the placeholders.
It doesn't solve the Android problem but it makes the input field usable.

In addition, cutting the value will erroneously remove an additional character:
Have a mask such as `111-11-1111` and enter `123-45-6789`. If you remove the `6` and leave the field, the cut will remove the ending character and will cause the input to be `123-45-_78_`. With the new change it will be `123-45-789_`.