Fix ref incompatibility with newer ember
ember-ref-modifier is unsupported in favor of ember-ref-bucket. This change is enough to get things working on ember 3.28 for me.

This change probably bumps the minimum supported ember version for this library to 3.20, since that is the oldest ember-ref-bucket supports.

Also, fellow app upgraders should take care that they aren't getting multiple conflicting copies of ember-modifier, because a newer version of ember-modifier will come as a dependency of ember-ref-bucket. In my case it was necessary to use yarn resolutions to put everything onto ember-modifier ^3.0.0.